Set _GNU_SOURCE in asyncsendto.c
Add #define on _GNU_SOURCE at the very beginning of asyncsendto.c file
as this file uses in6_pktinfo which is protected by __USE_GNU in
libc/inet/netinet/in.h. Currently, this flag is set by the Makefile in
CFLAGS however CFLAGS could be overwritten by the build system.

Signed-off-by: Fabrice Fontaine <fabrice.fontaine@orange.com>