Fix incorrect addition of rule to PREROUTING mangle chain
In the current iptables init script, the mangle chain has a rule added to jump to the MINIUPNPD table in the mangle chain added to PREROUTING. However, if you look at `iptables_removeall.sh`, it tries to remove the rule from the `FORWARD` chain instead. This changes the init script to add the jump rule to the FORWARD table instead of the PREROUTING table.