miniunpnpd: Unify naming scheme and rename to ext_allow_private_ipv4 for ignore_private_ip_check
Currently we are using a mix of "ignore private ip" and "ignore private
ip check" in code. Let's unify them to "ext allow private ipv4". We
always use private ip in most cases for internal interfaces, and it
would be confusing if we don't use ext prefix.

Besides, full-cone NAT is renamed as endpoint-independent NAT in newer
IETF RFC document. Add this name and a source of it to miniupnpd.conf.

Signed-off-by: Shengyu Qu <wiagn233@outlook.com>