miniupnpc: Use poll() where possible + Fix FD_SET crash bug
This commit modifies the function connecthostport in connecthostport.c
in order to prefer poll() over select(), on platforms that have it (such
as POSIX systems). For platforms that lack it, the fallback is select()
(such as on Windows or AmigaOS).

The reason why poll() is preferable here is that it does not suffer from
the limitations of select (see: man 2 select, BUGS section), where if
the file descriptor is FD_SETSIZE or greater, you cannot select() on it.

Additionally, this commit fixes #857, whereby the process could crash
due to unchecked use of FD_SET when the fd is outside FD_SETSIZE.